Оптимизация типов, алгоритма записи в битмап

+++ src/LinAlg/Vector2.cs
@@ -3,9 +3,9 @@ namespace RainRTX_Demo
 {
     public struct Vector2
     {
-        public double x, y;
+        public float x, y;
 
-        public Vector2 (double x, double y)
+        public Vector2 (float x, float y)
         {
             this.x = x;
             this.y = y;
